Allow ALAC to fill SampleEntry with Codec Parameters from decoded data
