+++ symphonia-common/src/apple/audio/mod.rs
@@ -1,0 +1,8 @@
+// Symphonia
+// Copyright (c) 2019-2025 The Project Symphonia Developers.
+//
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
+pub mod alac;
